segmentation/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy<2
transformers
torch
opencv-python-headless
accelerate
